chore: move filterURLsByProtocol function to utils.go
