+++ internal/pkg/postprocessor/outlinks.go
@@ -2,7 +2,6 @@ package postprocessor
 
 import (
 	"io"
-	"regexp"
 	"strings"
 
 	"github.com/internetarchive/Zeno/internal/pkg/config"
@@ -13,10 +12,6 @@ import (
 	"github.com/internetarchive/Zeno/internal/pkg/postprocessor/sitespecific/truthsocial"
 	"github.com/internetarchive/Zeno/internal/pkg/utils"
 	"github.com/internetarchive/Zeno/pkg/models"
-)
-
-var (
-	skipProtocolsRe = regexp.MustCompile(`(?i)^(data|file|javascript|mailto|sms|tel):`)
 )
 
 func extractOutlinks(item *models.Item) (outlinks []*models.URL, err error) {
@@ -131,16 +126,6 @@ func extractLinksFromPage(URL *models.URL) (links []*models.URL) {
 	}
 
 	return links
-}
-
-func filterURLsByProtocol(links []*models.URL) []*models.URL {
-	var filtered []*models.URL
-	for _, link := range links {
-		if !skipProtocolsRe.MatchString(link.Raw) {
-			filtered = append(filtered, link)
-		}
-	}
-	return filtered
 }
 
 func shouldExtractOutlinks(item *models.Item) bool {
